pyproject.toml
[project]
name = "neobridge"
version = "0.2.0"
description = "Lets you control neopixels from your PC using a board."
authors = [
    {name = "Zayne Marsh", email = "saynemarsh9@gmail.com"},
]
dependencies = [
    "pyserial>=3.5",
    "supervisor>=4.2.5",
    "pyautogui>=0.9.54",
    "typer>=0.12.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
